+++ crates/hpd-core/src/transition.rs
@@ -1,10 +1,10 @@
 use hpd_capabilities::power::PowerEnvelopeTarget;
-use hpd_capabilities::profile::{ProfileName, SystemPreset};
+use hpd_capabilities::profile::{ProfileName, TdpPreset};
 
 /// Represents any external event who is trying to alter the state
 #[derive(Debug, Clone)]
 pub enum Transition {
-    SetPreset(SystemPreset),
+    SetPreset(TdpPreset),
     SetSpl(u32),
     SetEnvelope(PowerEnvelopeTarget),
     SetProfile(ProfileName),